Trait iostream::io::Stream

pub trait Stream {
    fn set_position(&mut self, _potion: u64) -> Result<(), String>;

    fn position(&mut self) -> u64;

    fn length(&mut self) -> u64;

    fn set_length(&mut self, _: u64) -> Result<(), String>;

    fn can_write(&mut self) -> bool;

    fn can_read(&mut self) -> bool;

    fn can_seek(&mut self) -> bool;

    fn read_all(&mut self, buf: &mut Vec<u8>) -> Result<usize, String>;

    fn read(
        &mut self, 
        buf: &mut [u8], 
        offset: usize, 
        count: usize
    ) -> Result<usize, String>;

    fn read_byte(&mut self) -> Result<u8, String>;

    fn seek(&mut self, offset: i64, origin: SeekOrigin) -> Result<u64, String>;

    fn write_all(&mut self, buf: &[u8]) -> Result<(), String>;

    fn write(
        &mut self, 
        buf: &[u8], 
        offset: usize, 
        count: usize
    ) -> Result<usize, String>;

    fn flush(&mut self) -> Result<(), String>;

    fn copy_to(&mut self, target: &mut Stream) -> Result<(), String> { ... }
}

Required Methods

fn set_position(&mut self, _potion: u64) -> Result<(), String>

gets or sets the position within the current stream

fn position(&mut self) -> u64

The current position within the stream.

fn length(&mut self) -> u64

A long value representing the length of the stream in bytes.

fn set_length(&mut self, _: u64) -> Result<(), String>

Sets the length of this stream to the given value.

fn can_write(&mut self) -> bool

gets a value indicating whether the current stream supports writing. true if the stream supports writing; otherwise, false.

fn can_read(&mut self) -> bool

gets a value indicating whether the current stream supports reading. true if the stream supports reading; otherwise, false.

fn can_seek(&mut self) -> bool

gets a value indicating whether the current stream supports seeking.

fn read_all(&mut self, buf: &mut Vec<u8>) -> Result<usize, String>

When overridden in a derived class, reads a sequence of bytes from the current stream and advances the position within the stream by the number of bytes read.

fn read(
&mut self,
buf: &mut [u8],
offset: usize,
count: usize
) -> Result<usize, String>

reads a sequence of bytes from the current stream and advances the position within the stream by the number of bytes read.

fn read_byte(&mut self) -> Result<u8, String>

Reads a byte from the stream and advances the position within the stream by one byte, or Err returns -1 if at the end of the stream.

fn seek(&mut self, offset: i64, origin: SeekOrigin) -> Result<u64, String>

sets the position within the current stream.

fn write_all(&mut self, buf: &[u8]) -> Result<(), String>

writes a sequence of all bytes to the current stream

fn write(
&mut self,
buf: &[u8],
offset: usize,
count: usize
) -> Result<usize, String>

writes a sequence of bytes to the current stream and advances the current position within this stream by the number of bytes written.

fn flush(&mut self) -> Result<(), String>

clears all buffers for this stream and causes any buffered data to be written to the underlying device.

Provided Methods

fn copy_to(&mut self, target: &mut Stream) -> Result<(), String>

copy current stream to target stream

  use iostream::io::*;
  use std::fs::File;

  let mut ms = MemoryStream::new();
  let data:[u8;50]=[54;50];
  ms.write(&data,0,50).unwrap();

  let mut fs=File::open_fs("c.data",FileMode::CreateNew,FileAccess::ReadWrite).unwrap();
  ms.copy_to(&mut fs);